move get email to own package doesnt seem right to be in the slack package
